+++ src/index.js
@@ -1,7 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App'
-import './style.css'
-ReactDOM.render(<>
-<App/>
-</>,document.getElementById('root'))
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter } from "react-router-dom";
+import App from "./App";
+
+const app = (
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+);
+ReactDOM.render(app, document.getElementById("root"));
+
+
+
